+++ src/v0/destinations/fb_custom_audience/transform.ts
@@ -69,6 +69,7 @@ const preparePayload = (
     isHashRequired,
     disableFormat,
     destinationId,
+    workspaceId,
   );
   return batchingWithPayloadSize(prepareFinalPayload, workspaceId);
 };